Split config and schema
so as not to compile a schema unless needed
